Add files via upload
Description
Dataset_1:
The dataset consists of transaction timestamps (in hours) for a sample of online banking transactions. The timestamps represent the time of day when the transactions occurred.

Dataset_2:
The dataset, encapsulated as a pandas DataFrame "trans_David", chronicles the transactional activities of an individual named David. A salient column, "channel_cd", signifies the payment channel employed by David for each transaction. The dataset encompasses 40 entries across 14 columns, with 'channel_cd' being the focal point for the derivation of the 'freq_channel' feature.